Render keyframes on client to allow code snippets in definition
You can now do:

    (defanim foo
      [  0 (str "background: " colors/text-blue ";")]
      [100 (str "background: " colors/text-purple ";")])

The drawback is that the keyframes are now calculated on the client, which means
a very slight performance hit - but this only happens at startup. The
performance hit might even be outweighed by the reduced number of bytes being
sent over the wire. (Since the frames data is lighter than the resulting css).